Implemented resolution of a human readable url to the corresponding ID.
It is not bug free, it crashes about the second level or so.
TIt starts looking from the point it found a valid ID.
Then it proceeds forwards one path segment at a time and tries to
resolve the next step in the path. If it fails, it aborts.
The 3 slotResolveId overloads exist only because signals and
slots cannot resolve superclass relationships.