CDS event variables are listened for, and the right path is updated using a hash to store mappings.
The updates are checked every time the event loop runs, the trick is to run them periodically, which I don't think is possible.

I have a query based idea wherein by invoking the slave with a KIO::stat and passing a parameter like ?checkupdates=1 the slave will just process updates rather than perform any IO related job. Then the user of the slave ( UpnpCollection ) could periodically poll the slave. A KIO::listRecursive with a similar query would only return the updated folders. It remains to be seen how intuitive this is