Separated slave into two threads.
The controlpoint is managed by one thread which now has its own event loop and can keep watching for updates and stuff and avoid the forced eventing of the original kioslave. Browsing the top level works, but resolvePathToObject has to be modified to work in the async manner without involving a event loop just to block so that the slave never blocks.

updateDeviceInfo on the other hand, blocks until it finds a device, since without the device the slave is useless, and so it is a valid requirement.