Move starting the thread to the ControlPointThread constructor.
The moveToThread call is important. It effectively isolates
the two event loops since the ControlPointThread object
is now owned by itself in terms of thread affinity.

This was the bug which caused a busy loop to block the main
thread itself. With this in place, the ControlPointThread keeps
listening for changes while the main thread can still invoke KIO
commands.