browsing is now done in parts.
Rather than asking the CDS for all entries at once, it is
done in groups of 30 items each. Apparently a complete dump causes some implementations to crash for lots of data. The final plan is to have a relatively intelligent 'scheduler' to select the right interval, but 30 will do for now.

Having a global variable is a workaround until Tuomo makes the HAsyncOp setUserData stuff shared, so that the slot can actually access the user data.

The method prototypes have been changed, since browseResult and co. will now need to know where to start from and which object to call the browse action on again.