Converted to passing the browsing status through HAsyncOp userData.
This makes quite a few changes in the slave internals in the sense of method declarations. Most importantly browseDevice with 6 arguments now takes a DIDL::Object rather than a string ID. This is to accomodate 'recursively' calling browseResolvedPath rather than browseDevice while still being able to set the BrowseCallInfo in browseDevice() since that is a lower level detail.