Unified much of the search/browse API.
All the middle-man stuff like invoking actions, checking for errors and emitting signals is now common. Everything is started by browseOrSearchObject(). it finally emits browseResult to which the interested slot should connect. At the moment createSearchListing() is not unified since it might require slightly different parsing owing to the need to map ids to paths, which is a slightly sticky problem