kioslave now works with multiple upnp devices at once.
The KIO system reuses a slave when possible. This means that a slave can be called to browse multiple devices at once, side by side.
The older slave did a costly rescan of the device every time it changed. In addition it cleared the object cache and all information about other devices. When the device reverted back to the first device, all this information had to be reacquired. In addition the scan was liable to fail if the device was already known.
The slave now maintains a hash of all devices it knows about, and wraps each UPnP MediaServer in a MediaServerDevice struct which contains some addition details alongwith the HDevice. When a device is 'known', it is not a simple matter of reassigning the current device to start browsing it. Since the cache is still active, no performance loss occurs. For normal use, the memory consumed should not be a problem. This allows the Amarok collection to now have two or more devices as collections with the UpnpBrowseCollection working fine.

Updates are disabled, and will be fixed when it is possible to resolve the source of the update