The search API can now resolve paths to search results.
If the resolvePath attribute is passed to a search, the results will be reported with the UDS_NAME of the result pointing to the relative path from the search query root. This is useful in filebrowsers and so on.
The resolution is NOT done by default, since it takes an inordinate amount of time and is not actually required by most applications. In fact the need of this code is debatable