Removed all dependency on Cagibi.
HUpnp works and it works well, besides Cagibi doesn't do detection well (fails for certain devices) and adds an extra step. The HControlPoint now does detection itself and has a 5 second timeout instead.

I do not plan to use Solid either in the slave since that is also just another extra step. Solid would be better of used in stuff like the Amarok Collection Factory where it can be used to monitor for devices and create collections as required