Fix stat cache issue.
If a UPnP entry is cached due to a path resolution
it has only the title and ID present, and not all the
extra meta-data. A stat call on a cached entry will only
return this, making even file retrieval impossible since
the entry doesn't have the HTTP fetch URL. Now after a
path resolution, a browse is forced on the item to fetch
all meta-data.